backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
anthropic==0.48.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
aiosqlite==0.21.0
pydantic-settings==2.1.0
greenlet==3.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0
